refactor: simplify REST recovery without weakening admission
